Fixed some Mockito Tests and strated adding more Unit Tests

+++ src/main/java/com/tsi/kirk/mcallister/microdemo/MicroDemoApplication.java
@@ -1,5 +1,6 @@
 package com.tsi.kirk.mcallister.microdemo;
 
+import com.tsi.kirk.mcallister.microdemo.exceptions.ObjectNotFoundException;
 import com.tsi.kirk.mcallister.microdemo.inventory.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
@@ -14,8 +15,9 @@ import java.util.Optional;
 @RequestMapping("/Home") //base URL
 public class MicroDemoApplication {
 
+	//Attributes -------------------------------------------------------------
 	@Autowired
-	private final ActorRepository actorRepo;
+	private ActorRepository actorRepo;
 	@Autowired
 	private FilmRepository filmRepo;
 	@Autowired
@@ -26,11 +28,10 @@ public class MicroDemoApplication {
 	private FilmTextRepository filmTextRepo;
 	@Autowired
 	private InventoryItemRepository inventItemRepo;
+	// -----------------------------------------------------------------------
 
-	public static void main(String[] args) {
-		SpringApplication.run(MicroDemoApplication.class, args);
-	}
-
+	//Constructors -----------------------------------------------------------
+	public MicroDemoApplication() {}
 	public MicroDemoApplication(ActorRepository actorRepo) {
 		this.actorRepo = actorRepo;
 	}
@@ -48,7 +49,17 @@ public class MicroDemoApplication {
 		this.filmTextRepo = filmTextRepo;
 		this.inventItemRepo = inventItemRepo;
 	}
+	// -----------------------------------------------------------------------
 
+	//Main Method ------------------------------------------------------------
+	public static void main(String[] args) {
+		SpringApplication.run(MicroDemoApplication.class, args);
+	}
+	// -----------------------------------------------------------------------
+
+	//Mappings (Inventory) ---------------------------------------------------
+
+	//Actor
 	@GetMapping("/All_Actors")
 	public @ResponseBody
 	Iterable<Actor> getAllActors() {
@@ -72,7 +83,9 @@ public class MicroDemoApplication {
 	@DeleteMapping("/Delete_Actor")
 	public @ResponseBody
 	void deleteActor(@RequestParam int id) {
-		actorRepo.deleteById(id);
+		if (actorRepo.findById(id).isPresent()) {
+			actorRepo.deleteById(id);
+		}
 	}
 
 	@PutMapping("/Update_Actor")
@@ -87,34 +100,73 @@ public class MicroDemoApplication {
 		}
 		actorRepo.save(a);
 	}
+	//
 
+	//Film
 	@GetMapping("/All_Films")
 	public @ResponseBody
 	Iterable<Film> getAllFilms() {
 		return filmRepo.findAll();
 	}
 
+	@GetMapping("/Get_Film")
+	public @ResponseBody
+	Film getFilm(@RequestParam int id) {
+		Optional<Film> foundFilm = filmRepo.findById(id);
+		return foundFilm.orElse(null);
+	}
+
+	@PostMapping("/Add_Film")
+	public @ResponseBody
+	void addFilm(@RequestParam String firstName, @RequestParam String lastName) {
+		//TODO switch to film
+	}
+
+	@DeleteMapping("/Delete_Film")
+	public @ResponseBody
+	void deleteFilm(@RequestParam int id) {
+		if (filmRepo.findById(id).isPresent()) {
+			filmRepo.deleteById(id);
+		}
+	}
+
+	@PutMapping("/Update_Film")
+	public @ResponseBody
+	void updateFilm(@RequestParam int id, @RequestParam String firstName, @RequestParam String lastName) {
+		//TODO switch to film
+	}
+	//
+
+	//Category
 	@GetMapping("/All_Categories")
 	public @ResponseBody
 	Iterable<Category> getAllCategories() {
 		return catRepo.findAll();
 	}
+	//
 
+	//Language
 	@GetMapping("/All_Languages")
 	public @ResponseBody
 	Iterable<Language> getAllLanguages() {
 		return langRepo.findAll();
 	}
+	//
 
+	//Film Text
 	@GetMapping("/All_Film_Texts")
 	public @ResponseBody
 	Iterable<FilmText> getAllFilmTexts() {
 		return filmTextRepo.findAll();
 	}
+	//
 
+	//Inventory
 	@GetMapping("/All_Inventory_Items")
 	public @ResponseBody
 	Iterable<InventoryItem> getAllInventoryItems() {
 		return inventItemRepo.findAll();
 	}
+	//
+	// -----------------------------------------------------------------------
 }
